move get_instance_handle_from_key to InstanceHandle::try_from_key
