+++ dds/src/implementation/actors/data_reader_actor.rs
@@ -40,11 +40,8 @@ use crate::{
             types::{ChangeKind, Guid, GuidPrefix, Locator, ENTITYID_UNKNOWN, GUID_UNKNOWN},
             writer_proxy::RtpsWriterProxy,
         },
-        rtps_udp_psm::udp_transport::UdpTransportWrite,
-        utils::{
-            actor::{Actor, ActorAddress},
-            instance_handle_from_key::get_instance_handle_from_key,
-        },
+        udp::udp_transport::UdpTransportWrite,
+        utils::actor::{Actor, ActorAddress},
     },
     infrastructure::{
         self,
@@ -1690,7 +1687,7 @@ impl DataReaderActor {
                 "Writer with matched topic and type found",
             );
             let instance_handle =
-                get_instance_handle_from_key(&discovered_writer_data.get_key().unwrap()).unwrap();
+                InstanceHandle::try_from_key(&discovered_writer_data.get_key().unwrap()).unwrap();
             let incompatible_qos_policy_list = self
                 .get_discovered_writer_incompatible_qos_policy_list(
                     &discovered_writer_data,
