+++ dds/src/implementation/mod.rs
@@ -3,5 +3,5 @@ pub mod data_representation_builtin_endpoints;
 pub mod data_representation_inline_qos;
 pub mod payload_serializer_deserializer;
 pub mod rtps;
-pub mod rtps_udp_psm;
+pub mod udp;
 pub mod utils;
