+++ dds/src/implementation/utils/mod.rs
@@ -1,2 +1,1 @@
 pub mod actor;
-pub mod instance_handle_from_key;
